pyproject.toml
[project]
name = "airline-passenger-satisfaction-random-forest"
version = "0.1.0"
description = ""
authors = [
    {name = "Jakub Chodziński",email = "xapixowy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
